+++ src/main/java/de/katzenpapst/amunra/client/renderer/RenderShuttle.java
@@ -51,8 +51,8 @@ public class RenderShuttle extends Render {
         if (var28 > 0.0F) {
             final float i = entity.getLaunched() ? (5 - MathHelper.floor_double(entity.timeUntilLaunch / 85)) / 10F
                     : 0.3F;
-            GL11.glRotatef((float) (MathHelper.sin(var28) * var28 * i * par9), 1.0F, 0.0F, 0.0F);
-            GL11.glRotatef((float) (MathHelper.sin(var28) * var28 * i * par9), 1.0F, 0.0F, 1.0F);
+            GL11.glRotatef(MathHelper.sin(var28) * var28 * i * par9, 1.0F, 0.0F, 0.0F);
+            GL11.glRotatef(MathHelper.sin(var28) * var28 * i * par9, 1.0F, 0.0F, 1.0F);
         }
 
         this.bindTexture(this.texture);
